MAINT: move normalization logic to _normalizations
