+++ torch_np/_binary_ufuncs.py
@@ -1,19 +1,14 @@
+from . import _helpers
 from ._detail import _binary_ufuncs
+from ._normalizations import ArrayLike, DTypeLike, SubokLike, normalizer
 
 __all__ = [
     name for name in dir(_binary_ufuncs) if not name.startswith("_") and name != "torch"
 ]
 
 
-from . import _helpers
-from ._detail import _util
-
-# TODO: consolidate normalizations
-from ._funcs import ArrayLike, DTypeLike, SubokLike, normalizer
-
-
 def deco_binary_ufunc(torch_func):
-    """Common infra for unary ufuncs.
+    """Common infra for binary ufuncs.
 
     Normalize arguments, sort out type casting, broadcasting and delegate to
     the pytorch functions for the actual work.
@@ -33,19 +28,9 @@ def deco_binary_ufunc(torch_func):
         signature=None,
         extobj=None,
     ):
-        if order != "K" or not where or signature or extobj:
-            raise NotImplementedError
-
-        # XXX: dtype=... parameter
-        if dtype is not None:
-            raise NotImplementedError
-
-        out_shape_dtype = None
-        if out is not None:
-            out_shape_dtype = (out.get().dtype, out.get().shape)
-
-        tensors = _util.cast_and_broadcast((x1, x2), out_shape_dtype, casting)
-
+        tensors = _helpers.ufunc_preprocess(
+            (x1, x2), out, where, casting, order, dtype, subok, signature, extobj
+        )
         result = torch_func(*tensors)
         return _helpers.result_or_out(result, out)
